Web: Move MMBWrapper out of pages folder
It will create a route if it is located there.
Use media queries for mmb layout
