+++ apps/web/pages/mmb/flight-services.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { FlightServices } from '@kiwicom/margarita-core';
 
-import MMBWrapper from './components/Wrapper';
+import MMBWrapper from '../../components/Wrapper';
 
 const FlightServicesScreen = () => (
   <MMBWrapper>
