+++ apps/web/pages/mmb/passenger-detail.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { PassengerDetail } from '@kiwicom/margarita-core';
 
-import MMBWrapper from './components/Wrapper';
+import MMBWrapper from '../../components/Wrapper';
 
 const PassengerDetailScreen = () => (
   <MMBWrapper>
