+++ apps/web/pages/mmb/share-booking.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { ShareBooking } from '@kiwicom/margarita-core';
 
-import MMBWrapper from './components/Wrapper';
+import MMBWrapper from '../../components/Wrapper';
 
 const ShareBookingScreen = () => (
   <MMBWrapper>
