+++ apps/web/pages/mmb/trip-services.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { TripServices } from '@kiwicom/margarita-core';
 
-import MMBWrapper from './components/Wrapper';
+import MMBWrapper from '../../components/Wrapper';
 
 const TripServicesScreen = () => (
   <MMBWrapper>
